requirements.txt
jira
PyGithub
google-generativeai
PyYAML
GitPython
black
flake8
pytest
fastapi
uvicorn